improved lf preview

+++ .local/scripts/lf/lf.sh
@@ -2,18 +2,20 @@
 set -e
 
 if [ -n "$DISPLAY" ]; then
-    export FIFO_UEBERZUG="${TMPDIR:-/tmp}/lf-ueberzug-$$"
+    if ! [ "$TERM" = "xterm-kitty" ]; then
+        export FIFO_UEBERZUG="${TMPDIR:-/tmp}/lf-ueberzug-$$"
 
-    cleanup()
-    {
-        exec 3>&-
-        rm "$FIFO_UEBERZUG"
-    }
+        cleanup()
+        {
+            exec 3>&-
+            rm "$FIFO_UEBERZUG"
+        }
 
-    mkfifo "$FIFO_UEBERZUG"
-    ueberzug layer -s <"$FIFO_UEBERZUG" &
-    exec 3>"$FIFO_UEBERZUG"
-    trap cleanup EXIT
+        mkfifo "$FIFO_UEBERZUG"
+        ueberzug layer -s <"$FIFO_UEBERZUG" &
+        exec 3>"$FIFO_UEBERZUG"
+        trap cleanup EXIT
+    fi
 
     if ! [ -d "$HOME/.cache/lf" ]; then
         mkdir -p "$HOME/.cache/lf"
